requirements.txt
numpy~=1.19.5
Flask~=2.0.1
sklearn~=0.0
scikit-learn~=0.24.2
python-mnist~=0.7